+++ backend/users/Touir.ExpensesManager.Users/Infrastructure/Contracts/IEmailService.cs
@@ -1,0 +1,14 @@
+namespace Touir.ExpensesManager.Users.Infrastructure.Contracts
+{
+    public interface IEmailService
+    {
+        bool SendEmail(
+            string? recipientTo = null,
+            string? recipientCC = null,
+            string? recipientBCC = null,
+            string? emailSubject = null,
+            string? emailBody = null,
+            bool isHTML = false,
+            ICollection<string>? attachments = null);
+    }
+}